refactor: seperate EditV2 defi from handleEdit
